refactor: 구간 추가 로직 책임 분리
